server/requirements.txt
fastapi
uvicorn
python-multipart
torch
transformers
numpy
pandas
umap-learn
hdbscan
biopython
matplotlib
seaborn
websockets